pyproject.toml
[tool.poetry]
name = "ocr-image"
version = "0.1.0"
description = ""
authors = ["gustavooliveira <gusatlan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^11.0.0"
pytesseract = "^0.3.13"
openai = "^1.57.0"
ollama = "^0.4.4"
google-genai = "^0.2.2"
google-generativeai = "^0.8.3"
paho-mqtt = "^2.1.0"
python-dotenv = "^1.0.1"
schedule = "^1.2.2"
opencv-python = "^4.10.0.84"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
